Debug Pawn Movement
So now, with the logic added to GameState, I am able to retrieve the avalible actions of a pawn, and select an action and execute it. For now its currently executing the first action it comes across ( Which is a standard movement in the first clockwise cardinal direction ), but it has the ability to attack as well!