Added Attack Success Chance
Pieces now when they attack will roll a die, to determine if their attack will succed or not. As of currently this value is only displayed in console.
Attacks can now be undone, and redone with the same seed roll, but attacking different pieces on the same seed.

Future iterations should add a proper display graphic for attack actions ( Aka a chance of success ) and a mark to indicate how many actions have been taken this turn, perhaps labeling if the attack was a failure