Added AI Debug mode, and an Action Display
Added an AI Debug mode, where the actions taken are shown on an improved and enlargened window.

The Action display shows you current and last move actions ( Aka your move and the moves of your oponent ). It displays the piece involved, its current location, and what the action was. In terms of actions it shows different icons depending on who attack and if the attack was successful